sam/python/s3_thumbnail/function/requirements.txt
requests==2.32.3
boto3==1.35.97
python-dotenv==1.0.1
